+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All important changes to `fiisoft-phinx` will be documented in this file
 
+## 1.1.0
+
+Added custom abstract class (PhinxAbstractMigration) as default base class for created migration classes.
+
+Added util class PhinxTemplatePath with nice method path(). 
+
 ## 1.0.0
 
 Initial version.